Some minor changes to be able to use SASS.
Only changed a couple of the style.css elements so that I can use SASS.
This is a fork of the brilliantly simple _s them by automattic.
I am using @import base.css to pull in my scss files. It's not elegant,
and uses 1 extra header request, but I think this is so insignificant
compared to the time I can save developing my wordpress themes.